+++ tests/TestCase/TestSuite/HelperTestCaseTest.php
@@ -29,8 +29,8 @@ class HelperTestCaseTest extends TestCase
      */
     public function testGetMagicMethod(): void
     {
-        $HelperTestCase = $this->getMockForAbstractClass(HelperTestCase::class, [], '', true, true, true, ['getOriginClassNameOrFail']);
-        $HelperTestCase->method('getOriginClassNameOrFail')->willReturn(HtmlHelper::class);
+        $HelperTestCase = $this->getMockForAbstractClass(HelperTestCase::class, [], '', true, true, true, ['getOriginClassName']);
+        $HelperTestCase->method('getOriginClassName')->willReturn(HtmlHelper::class);
         $this->assertInstanceOf(HtmlHelper::class, $HelperTestCase->Helper);
     }
 }