+++ tests/test_app/TestApp/Test/TestCase/BadTestClass.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace App\Test\TestCase;
+
+use MeTools\TestSuite\TestCase;
+
+class BadTestClass extends TestCase
+{
+}