+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * `RunAllCommand::$questions` requires a boolean for the `default` value (and no longer `Y` or `N` strings), while the
   `command` value requires an instantiated command (and no longer a class-string). This simplifies coding and testing;
 * `Command` class has been moved from `MeTools\Console` to `MeTools\Command`;
+* `MockTrait::getOriginClassNameOrFail()` become `getOriginClassName()`, replacing the latter;
 * added `CommandTestCase` to test commands. It provides `$Command` property through the `__get()` magic method;
 * `TestCase` no longer always uses the `ReflectionTrait`;
 * `getMockForComponent()` and `getMockForController()` methods provided by `MockTrait` and `assertExitWithError()` and